Refactored Resume Editor
-71% file size reduction
-Extracted components
