+++ src/components/ResumeEditor/index.ts
@@ -4,6 +4,25 @@ export { ProfessionalSummarySection } from './components/sections/ProfessionalSu
 export { TechnicalSkillsSection } from './components/sections/TechnicalSkillsSection';
 export { WorkExperienceSection } from './components/sections/WorkExperienceSection';
 export { EducationSection } from './components/sections/EducationSection';
+export { CoursesSection } from './components/sections/CoursesSection';
+export { InterestsSection } from './components/sections/InterestsSection';
+export { ProjectsSection } from './components/sections/ProjectsSection';
+export { LanguagesSection } from './components/sections/LanguagesSection';
+export { PublicationsSection } from './components/sections/PublicationsSection';
+export { AwardsSection } from './components/sections/AwardsSection';
+export { VolunteerExperienceSection } from './components/sections/VolunteerExperienceSection';
+export { ReferencesSection } from './components/sections/ReferencesSection';
+export { ExportPanel } from './components/ExportPanel';
+export { DatePicker } from './components/DatePicker';
+export { EditResumeInfoModal } from './components/EditResumeInfoModal';
+export { LayoutModal } from './components/LayoutModal';
+export { AddSectionPopup } from './components/AddSectionPopup';
+export { DeleteConfirmationDialog } from './components/DeleteConfirmationDialog';
+export { FloatingActionButton } from './components/FloatingActionButton';
+export { MainContentArea } from './components/MainContentArea';
+export { LoadingComponent } from './components/LoadingComponent';
+export { AlertMessages } from './components/AlertMessages';
+export { createSectionComponents } from './components/SectionComponents';
 export { useResumeData } from './hooks/useResumeData';
 export { useExportSettings } from './hooks/useExportSettings';
 export { formatPhoneNumber, formatDate, isDateInFuture } from './utils/dateUtils';
